+++ Source/sram.cpp
@@ -13,7 +13,7 @@ using namespace std;
 
 Sram::Sram(void) {
 
-    this->ram = (int*) malloc(SRAM_SIZE * sizeof(int));
+    this->ram = (int8_t*) malloc(SRAM_SIZE * sizeof(int));
 }
 
 Sram::~Sram() {
